requirements.txt
-i https://pypi.org/simple
pillow==11.0.0; python_version >= '3.9'
